requirements.txt
requests
mnemonic
bip32utils
eth-account
python-telegram-bot